fix <back> <set_username>: login is now separate
login and nickname are two separate entities now.
login is attributed by 42 login and will never change.
nickname starts as empty value that is changed when the user changes his
nickname with set_username.
If the front wants the names to display, if the user has not set any
nickname then the login will appear, otherwise it will be the nickname.

Naming convention will be changed soon.